Initial support for Visual Studio / vcpkg

+++ libs/core/src/Break.cc
@@ -126,7 +126,7 @@ Break::get_timer_remaining()
 {
   if (timer->is_limit_enabled())
     {
-      int remaining = timer->get_limit() - timer->get_elapsed_time();
+      int64_t remaining = timer->get_limit() - timer->get_elapsed_time();
       return remaining >= 0 ? remaining : 0;
     }
   return -1;
